Refactored and moved various files
